requirements.txt
fastapi[all]
uvicorn
httpx
python-dotenv
aiofiles
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
aiosqlite
slowapi
pydantic[email]
fastapi-mail
pytest
pytest-asyncio
coverage
pandas
reportlab
python-multipart
jinja2
bcrypt
